+++ src/utils/genRandomNumber.js
@@ -1,4 +1,4 @@
-import { MIN_NUMBER, MAX_NUMBER } from '../data/constatns.js';
+import { MIN_NUMBER, MAX_NUMBER } from './constants.js';
 
 const genRandomNumber = () => {
   return MissionUtils.Random.pickNumberInRange(MIN_NUMBER, MAX_NUMBER);
